add API for easier duplicate pages handling
closes #68
